requirements.txt
# Core dependencies
python-dotenv==1.0.0
google-genai

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# PDF generation (optional)
reportlab==4.0.7

# Testing
pytest==7.4.3
httpx==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openpyxl==3.1.2

# Database & Auth added
sqlalchemy==2.0.23
pymysql==1.1.0
